Remove unused controllers and associated tests (#268)
* Remove AccountController and its associated tests

* Remove TransactionController, related DTOs, interface, and tests

* Remove `UserController` and related classes, interfaces, and tests
