--- src/main/java/info/mackiewicz/bankapp/core/transaction/model/dto/CreateTransactionRequest.java
@@ -1,13 +1,0 @@
-package info.mackiewicz.bankapp.core.transaction.model.dto;
-
-import lombok.Data;
-
-import java.math.BigDecimal;
-@Data
-public class CreateTransactionRequest {
-    private Integer sourceAccountId;
-    private Integer destinationAccountId;
-    private BigDecimal amount;
-    private String type;
-    private String title;
-}
