+++ src/main/java/info/mackiewicz/bankapp/presentation/auth/service/UserRegistrationService.java
@@ -4,8 +4,8 @@ import info.mackiewicz.bankapp.core.account.model.Account;
 import info.mackiewicz.bankapp.core.account.service.AccountService;
 import info.mackiewicz.bankapp.core.transaction.model.Transaction;
 import info.mackiewicz.bankapp.core.transaction.service.TransactionService;
-import info.mackiewicz.bankapp.core.user.UserMapper;
 import info.mackiewicz.bankapp.core.user.model.User;
+import info.mackiewicz.bankapp.core.user.model.dto.UserMapper;
 import info.mackiewicz.bankapp.core.user.service.UserService;
 import info.mackiewicz.bankapp.presentation.auth.dto.UserRegistrationRequest;
 import info.mackiewicz.bankapp.system.notification.email.EmailService;
