+++ src/test/java/info/mackiewicz/bankapp/core/user/UserMapperTest.java
@@ -1,7 +1,7 @@
 package info.mackiewicz.bankapp.core.user;
 
 import info.mackiewicz.bankapp.core.user.model.User;
-import info.mackiewicz.bankapp.core.user.model.dto.UpdateUserRequest;
+import info.mackiewicz.bankapp.core.user.model.dto.UserMapper;
 import info.mackiewicz.bankapp.core.user.model.vo.PhoneNumber;
 import info.mackiewicz.bankapp.presentation.auth.dto.UserRegistrationRequest;
 import org.junit.jupiter.api.BeforeEach;
@@ -41,40 +41,5 @@ class UserMapperTest {
         // then
         assertNotNull(user);
         assertEquals(new PhoneNumber(TEST_PHONE), user.getPhoneNumber());
-    }
-
-    @Test
-    void updateUserFromRequest_shouldUpdatePhoneNumberCorrectly() {
-        // given
-        User existingUser = new User();
-        existingUser.setPhoneNumber(new PhoneNumber("+48111111111"));
-        
-        UpdateUserRequest request = new UpdateUserRequest();
-        request.setPhoneNumber(TEST_PHONE);
-
-        // when
-        User updatedUser = userMapper.updateUserFromRequest(existingUser, request);
-
-        // then
-        assertNotNull(updatedUser);
-        assertEquals(new PhoneNumber(TEST_PHONE), updatedUser.getPhoneNumber());
-    }
-
-    @Test
-    void updateUserFromRequest_shouldNotUpdatePhoneNumberWhenNull() {
-        // given
-        String originalPhone = "+48111111111";
-        User existingUser = new User();
-        existingUser.setPhoneNumber(new PhoneNumber(originalPhone));
-        
-        UpdateUserRequest request = new UpdateUserRequest();
-        request.setPhoneNumber(null);
-
-        // when
-        User updatedUser = userMapper.updateUserFromRequest(existingUser, request);
-
-        // then
-        assertNotNull(updatedUser);
-        assertEquals(new PhoneNumber(originalPhone), updatedUser.getPhoneNumber());
     }
 }
