+++ src/test/java/info/mackiewicz/bankapp/presentation/service/UserRegistrationServiceTest.java
@@ -6,8 +6,8 @@ import info.mackiewicz.bankapp.core.account.service.AccountService;
 import info.mackiewicz.bankapp.core.transaction.model.Transaction;
 import info.mackiewicz.bankapp.core.transaction.model.builder.TransferBuilder;
 import info.mackiewicz.bankapp.core.transaction.service.TransactionService;
-import info.mackiewicz.bankapp.core.user.UserMapper;
 import info.mackiewicz.bankapp.core.user.model.User;
+import info.mackiewicz.bankapp.core.user.model.dto.UserMapper;
 import info.mackiewicz.bankapp.core.user.model.vo.EmailAddress;
 import info.mackiewicz.bankapp.core.user.service.UserService;
 import info.mackiewicz.bankapp.presentation.auth.dto.UserRegistrationRequest;
